+++ src/main/java/com/github/heikyudev/maestrocervecero/service/response_dto/receta/VersionRecetaResponseDTO.java
@@ -31,7 +31,7 @@ public class VersionRecetaResponseDTO {
     /**
      * Relación de empaste de la receta.
      */
-    private Integer relacionDeEmpaste;
+    private Double relacionDeEmpaste;
 
     /**
      * Densidad original objetivo de la receta.
